Adicionada modificações no AuthResource

+++ src/main/java/com/rodrigor/meat/security/UserSS.java
@@ -13,7 +13,7 @@ import com.rodrigor.meat.domain.enums.Profile;
 public class UserSS implements UserDetails{
 	private static final long serialVersionUID = 1L;
 	
-	private Integer id;
+	private Long id;
 	private String email, password;
 	
 	private Collection<? extends GrantedAuthority> authorities;
@@ -22,7 +22,7 @@ public class UserSS implements UserDetails{
 		
 	}
 	
-	public UserSS(Integer id, String email, String senha, Set<Profile> profile) {
+	public UserSS(Long id, String email, String senha, Set<Profile> profile) {
 		super();
 		this.id = id;
 		this.email = email;
@@ -39,7 +39,7 @@ public class UserSS implements UserDetails{
 		return getAuthorities().contains(new SimpleGrantedAuthority(profile.getDescription()));
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
